Moves context-related code into a single module called 'context'. Includes code for building arrays from contexts
